rename: model.py -> policy.py
