+++ src/model/__init__.py
@@ -1,2 +1,2 @@
-from .model import Policy
+from .policy import Policy
 from .critic import Critic
